ai-adk-mcp-cloud-run-flightbooking/mcp_server/requirements.txt
mcp[cli]
fastmcp
google-cloud-bigquery